misc!: remove `MapperBuilderConfiguratorAttribute` support
The feature that allowed attributes to configure a `TreeMapper` directly
during the injection has been removed.

After some thoughts, this implementation was a bad idea that led to more
complexity in the bundle code base, for something that should anyway be
done differently.

There will be no replacement for this feature and code that used it
should instead either inject an instance of `MapperBuilder` or use the
`MapperBuilderConfigurator` interface.